+++ CHANGELOG.md
@@ -11,6 +11,8 @@
   see [#393](https://github.com/cadeyrn/newtaboverride/issues/393)
 - Removed the dedicated button for changing the settings shortcut from the settings page. Shortcut settings remain
   available through Firefox's add-ons manager, see [#394](https://github.com/cadeyrn/newtaboverride/issues/394)
+- Added a theme switcher to the settings page for choosing light mode, dark mode, or automatic appearance based
+  on the system preference, see [#395](https://github.com/cadeyrn/newtaboverride/issues/395)
 
 #### Bugfixes